Include all sets search
Added the function to read from the initial page
(https://scryfall.com/sets) and go through the page creating a file with
each card of each set of Magic the Gathering's history.

There are still some problems as some times the programs crashes and
can't continue to create the files. I'm working to solve it.